+++ lib/jei/relationship.rb
@@ -9,6 +9,8 @@ module Jei
     # @option options [Boolean] :no_data Whether to exclude data from the
     #   relationship.
     # @option options [Proc] :links A Proc that evaulates to a list of links.
+    # @option options [Class] :serializer Override the default serializer used
+    #   for related resources.
     def initialize(name, value = name, options = {})
       super(name, value)
       @options = options